requirements.txt
# Flask and Web Framework
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.1

# Firebase
firebase-admin>=6.2.0
pyrebase4>=4.7.1

# NLP and Text Processing
spacy>=3.8.0
nltk>=3.8.1
keybert>=0.8.3
sentence-transformers>=2.2.2
rake-nltk>=1.0.6

# PDF Processing
pdfplumber>=0.10.3
PyMuPDF>=1.23.6
python-docx>=1.1.0

# Machine Learning and Transformers
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0

# Knowledge Graph
networkx>=3.2.1
pyvis>=0.3.2

# Multi-Armed Bandit
numpy>=2.0.0
scipy>=1.11.4
scikit-learn>=1.3.2

# Utilities
pandas>=2.1.4
tqdm>=4.66.1
requests>=2.31.0

# Testing
pytest>=7.4.3
